Refactor to follow conventions
